requirements.txt
Flask==2.2.5
pymysql
sqlalchemy
flask_sqlalchemy
python-dotenv

